Gov Pay integration
IVORY-305: User can pay

Added call to Gov Pay to allow payment for applications.

New routes:
service-complete: The last page in the flow that the user sees if their payment was successful
make-payment: An intermediate page that is used to send payment information over to Gov Pay.

If a payment is successful, the user ends up at service-complete
If the payment fails, the user goes back to make-payment, and hence goes back to Gov Pay
If the user cancels the payment in Gov Pay they are returned to check-your-answers

Payment calculation - Added to the 'what type of item is it' page so that we know how much to charge the applicant. Two payment bands have been introduced (A & B) that are held in config.

Reference numbers - A 6 character random string is being generated to be used as the reference number for payments, which is also shown in the service-complete page.

Includes fix to cookie config, as previously the config was not being applied.

Added ability to hide the back link for pages where this is not applicable. Back link is hidden if 'hideBackLink' boolean is passed into the page context.

Note: Dummy Gov UK card numbers can be found here:

https://docs.payments.service.gov.uk/testing_govuk_pay/#mock-card-numbers

Also: Added missing unit tests for Applicant contact details screen