Fixes for 327 and 329 (refinement of address matching)
IVORY-327: User goes to choose address if no match on name/number
IVORY-329: How do we adapt / refine search for house name/number

This is an update to our address filtering. The address API does not allow us to pass in building name/number, we filter for this on our side.

Previously we were filtering based on exact match with either BuildingNumber, BuildingName or SubBuildingName.

Now, if there are no matches but there are results for a postcode, the user is sent to the 'Choose address' screen.

We now do partial matches, so if the user enters the full first line of their address in the building name/number field (e.g. 123 Some Street) then we handle that by matching on the corresponding address.

Also, if the user only enters part of the building name (e.g. 'barn' rather than 'The Barn') then we match on that too.

In an effort to be as flexible as possible with the formatting, all non-alphanumeric characters and whitespace is disregarded when looking for matches. Matches are case-insensitive.