Describe the item page (S10 and S2 descriptions)
IVORY-352: Describe the item page

This change adds the content for the /describe-the-item page.

Included is a function called _addRedisDataToContext. This is responsible for populating the page on GET if we are returning there later, e.g. from Check Your Answers.

A similar fix has been carried out on the following pages:

- what-type-of-item-is-it
- who-owns-the-item

Also included is an increase to the request timeout on the Upload page to 2 minutes to fix a problem that occurred when uploading large files.