Suppress logging of 401 errors
When basic authentication is turned on, the server writes information to the console whenever the user is prompted for their username and password. These error logs are unnecessary and misleading, as the authentication server error was handled normally, therefore it makes sense to suppress this error logging.

Another thing that was added was a "catch all" for 404 errors (page not found), because otherwise if basic authentication is turned on, the login prompt appears and the user presses 'cancel' then a 404 error occurs. Adding this catch-all prevents that from happening.