Features 365 and 366 - Handle user session timing out or user joining part-way through the flow
IVORY-365 Handle user not having a valid session cookie
(also incorporates IVORY-366 Handle user joining part-way through the flow)

This has been implemented by creating a new 'onPreResponse' handler that checks the session cookie on all GETs and POSTs. If the user does not have a valid cookie then they are automatically redirected to the following new route:

/errors/session-timed-out

NB - The copy for this page is still TBC so the provided page in this commit is just a placeholder.

The only pages excluded from this cookie check are:

/ (i.e. home)
/errors/* (e.g. /errors/page-not-found)

In order to check the cookie a new CookieService has been developed.

As a result of this change, almost all unit tests needed to be updated to mock the CookieService, as all the pages are now expecting a valid cookie.

The default cookie timeout is new config, set to 24 hours in milliseconds.

This ticket also included a related copy change on the /can-continue page.