Upload additional evidence for S2 (high value) applications
IVORY-367: Upload additional evidence for S2
IVORY-391: Update front office flow

This change adds the front end screens for the uploading of additional evidence for S2 (high value) applications. These screens are based on the existing image upload screens, but allow the user to select up to 6 PDF or MS Word documents.

The selected documents are added to the PowerApps record for the application once it has been created. Separate PowerApps WebAPI calls are needed for each document that has been uploaded by the user.

This change also includes a change to the front office flow as per ticket 391.