Version 11.1.0 (#114)
* Upload additional evidence for S2 (high value) applications

IVORY-367: Upload additional evidence for S2
IVORY-391: Update front office flow

This change adds the front end screens for the uploading of additional evidence for S2 (high value) applications. These screens are based on the existing image upload screens, but allow the user to select up to 6 PDF or MS Word documents.

The selected documents are added to the PowerApps record for the application once it has been created. Separate PowerApps WebAPI calls are needed for each document that has been uploaded by the user.

This change also includes a change to the front office flow as per ticket 391.

* Upgraded dependencies, fixed some SonarCloud issues

Upgraded dependencies, fixed some SonarCloud issues.

* Updated copy on /can-continue page

IVORY-416: You can continue page

Updated copy on /can-continue page

* Update to copy on cannot-continue page

IVORY-414: Update to 'cannot continue' page

Copy update on the following page: /eligibility-checker/cannot-continue

* Updated copy

IVORY-415: Update to the 'What type of item is it' screen

Update to the copy on the /what-type-of-item-is-it screen.

* Incremented version number

* Incremented version number

* Various minor screen amendments

IVORY-400: Design QA of the Dev environment

This change includes fixes to various screens where the design had moved on but the development had not been carried out.

* Reinstate page missing from the flow

IVORY-421: Missing CD72 - Do you want to add any documents to support your application?

The page flow was incorrect and the screen for CD72 was not included in the flow.

* Feature/ivory 396 confirmation page (#111)

* updated packages

* Updated page content

* created unit tests

* Fixed vulnerability in node_modules/tar

* Remove get-record route

Removed POC get-record route prior to Private Beta release.

Co-authored-by: Jason Claxton <30830544+Jozzey@users.noreply.github.com>