Check your answers MVS
IVORY-399: Check your answers MVS

This change adds a new version of the Check Your Answers (CYA) page. For now, the Change links take you back to a particular page for editing, but once you are there you need to go all the way back through subsequent screens to get back to the CYA page. This may be amended in a later version.

Also included in this change is a new way of mocking the data from the Redis cache. Now, the mock function returns the data corresponding to a particular Redis key, rather than having to have multiple Redis mock return values in a specific order.

A circular page flow bug was also found during development and has been fixed.