S10 registration confirmation emails
IVORY-433: S10 registration confirmation emails

This change adds integration with GOV Notify in order to send confirmation emails to Section 10 applicants.

The emails are sent once payment has successfully been made, on the /service-complete page. There is a check to ensure that if the page is refreshed, the email is only sent on the first page visit, in order to prevent multiple emails from being sent.

This change also fixes a contact details bug introduced in the last commit.