Various code and unit test refactoring
IVORY-445: Unit tests don't work without internet access

Fixed problem of unit tests not running when not connected to the internet due to a recently added Google Analytics plugin. A new AnalyticsService has been added to allow the call to the plugin to be stubbed out.

Refactored the _getContext methods in route files so that this is only called once per GET or POST.

Added new TestHelper.createServer() method that now mocks out the Google Analytics library('@defra/hapi-gapi') before creating the test server, to save having to do this in all unit test files.