Fixes for error when uploading large documents and images
IVORY-419: Error when uploading large documents and images

Initially, documents and images of up to 30MB were allowed. However, this caused a problem when the application was deployed into the Azure infrastructure, due to timeout issues. In order to solve this issue, it has been agreed that the file size limit can be reduced to 10MB.