Fixed file upload problem causing "file too large" error messages not to appear
IVORY-419: Image and file uploads - "file too large" error message not appearing

While investigating IVORY-419 it was noticed that the "file too large" error message was no longer appearing. This was because the way the boolean values in Redis are saved has recently changed. This is now fixed - the "file too large" error does now appear, but it is unlikely that this has fixed the actual problem reported in IVORY-419, but further investigation of that issue will now be possible.