Automatic email changes
IVORY-580: S10 Confirmation email content change
IVORY-581: 3rd party S2 journeys: send emails to owner as well as applicant

Added 2 new emails for S2 journeys. Refactored the environment variables and the code responsible for the sending of emails to cope with the addition of the extra 2 emails (now a total of 6).

Also made the minor content change for 580 - on one of the S10 emails there is now a piece of dynamic copy that only appears in the email if the exemption type is "MUSEUM". Note that this is a breaking change which has needed a new email template:

Section 10 - Applicant confirmation (v5)
The original email template can be removed once this change has propagated into all environments:

Section 10 - Applicant confirmation