Lock down functionality in the back office
IVORY-490: Lock down functionality in the back office

A new Ivory Superuser role has been created. Fields within the back office system are now enabled / disabled depending on whether or not the user is an Ivory User or Ivory Superuser and also whether or not the record was created manually (i.e. directly in the PowerApps application) or automatically (i.e. via the front end service).