Moved generation of submission reference number to earlier in the flow
IVORY-600: Submission reference is being re-created each time a payment attempt is made

The submission reference is now created at the beginning of the journey (the / page) at the same time as the session cookie is created. It is then stored in Redis and obtained from there when each payment attempt is made.