Fixed problem with fields only containing numeric characters
IVORY-622: Error after payment page: 3rd party RMI application

There was an issue with the Redis Service that was causing a service error page to be displayed.

In RedisService.get we were returning a string from the Redis cache and then we are trying to Json.parse it back into an object, but if the value was a number, e.g. ‘123’, and not part of a JSON object, then the Json.parse function was converting it from the string into a number 123, which then bombed out the Web Api call on that field.

This has been changed so that now the Redis Service only parses the strings coming out of the cache if they are JSON objects.