Emails not sent to third party business owner
IVORY-618: Emails not sent to third party business owner

Section 2 and Section 10 emails were not being sent to the owner when the owner is a business and the application is being made by a third party (i.e. not the applicant).

The reason is that in this situation the fullName field is blank, instead we have to use the businessName field.