Added values to Redis that are bypassed when coming in via the new /use-checker route
IVORY-629: Using Direct Eligibility checker link - 500 error after declaration page

Added values to Redis that are bypassed when coming in via the new /use-checker route, instead of coming in via /eligibility-checker/how-certain.

Method:

Look up the submission reference - this may have already been set in the previous screen. If not, i.e. the user came in via the new /use-checker route, then generate the submission reference and store in Redis.

Set the used-checker flag to true - this will not have been set on the previous screen if the user has come in via the new /use-checker route.