Make Redis Time to Live (TTL) configurable
IVORY-630: Make Redis cache timeout configurable

This change utilises the existing COOKIE_TIMEOUT environment variable, which currently controls the user session timeout. This is now used to also set the Redis Time to Live (TTL) as well as to set some dynamic copy on the /can-continue page.

This will allow us to expire data from the Redis cache more quickly if we find that it is filling up with data.