Provide performance test mode
IVORY-631: Provide a performance test mode

This change adds a new PERFORMANCE_TEST_MODE environment variable which, when set, causes the flow to go straight from the /make-payment route to the /service-complete route. This is for performance test purposes only and is used to bypass the following parts of the service:

- Making a Gov Pay payment
- Sending Gov Notify confirmation emails
- Saving a record in the Dataverse

In the longer term we should look to implement stubs for Gov Pay, Gov Notify and the Dataverse, but this change will allow us to complete the user journeys in the performance test mode and allow us to demonstrate that when the journeys are completed, the corresponding user session data is removed.