Blob storage of documents and images
IVORY-632: Use Azure blob storage for temporary storage of documents and images instead of Redis

This change moves the storage of documents and images from Redis into a new Azure storage account.

The in-memory Redis cache is only recommended for use to store objects up to 100kb in size. The Ivory service allows users to upload documents and images that are up to 10MB in size. Using Redis to store objects of this size is costly to scale up to support the expected number of users and has potential performance problems for the storage and retrieval of this size of object.

A new storage account has been set up that allows the storage and retrieval of files as BLOBs (binary large objects). Azure blob storage is much cheaper than Redis storageAzure Blob Service.

A new AzureBlobStorage service has been introduced with is used to connect to the Azure storage account.

The following new environment variables are required:

AZURE_STORAGE_ACCOUNT
AZURE_STORAGE_ACCOUNT_KEY

For reference, Redis best practices:
https://docs.microsoft.com/en-us/azure/azure-cache-for-redis/cache-best-practices-development

Microsoft documentation on potential Redis performance issues:
https://techcommunity.microsoft.com/t5/azure-paas-blog/azure-redis-timeouts-client-side-issues/ba-p/2022203