Techdebt/ivory 504 upgrade npm dependences (#273)
All dependancies updated to latest versions except for node-fetch. The latest version of node-fetch is an ESM-only module which is not currently supported by our application. Advice from the creators of node-fetch is to remain on v2:

This module was converted to be a ESM only package in version 3.0.0-beta.10. node-fetch is an ESM-only module - you are not able to import it with require. We recommend you stay on v2 which is built with CommonJS unless you use ESM yourself. We will continue to publish critical bug fixes for it.